website creation and deletion added to controller

+++ Hipstapaper/Packages/Datum2/Sources/Datum2/Controllers/CD_Controller.swift
@@ -33,191 +33,50 @@ extension CD_Controller: Controller {
     internal var ENVIRONMENTONLY_managedObjectContext: NSManagedObjectContext {
         self.container.viewContext
     }
-    
-    /*
 
     // MARK: Website CRUD
 
-    func createWebsite(_ raw: AnyWebsite.Raw) -> Result<AnyElementObserver<AnyWebsite>, Error> {
+    internal func createWebsite(_ input: Website?) -> Result<Website.Ident, Error> {
         assert(Thread.isMainThread)
-
         let context = self.container.viewContext
-        let website = CD_Website(context: context)
-        if let title = raw.title {
-            website.cd_title = title
-        }
-        if let originalURL = raw.originalURL {
-            website.cd_originalURL = originalURL
-        }
-        if let resolvedURL = raw.resolvedURL {
-            website.cd_resolvedURL = resolvedURL
-        }
-        if let isArchived = raw.isArchived {
-            website.cd_isArchived = isArchived
-        }
-        if let thumbnail = raw.thumbnail {
-            website.cd_thumbnail = thumbnail
+        let cd_website = CD_Website(context: context)
+        if let input = input {
+            cd_website.cd_title       = input.title
+            cd_website.cd_originalURL = input.originalURL
+            cd_website.cd_resolvedURL = input.resolvedURL
+            cd_website.cd_thumbnail   = input.thumbnail
+            cd_website.cd_dateCreated = input.dateCreated
         }
         return context.datum_save().map {
-            AnyElementObserver(ManagedObjectElementObserver(website, { AnyWebsite($0) }))
-        }
-    }
-
-    func readWebsites(query: Query) -> Result<AnyListObserver<AnyRandomAccessCollection<AnyElementObserver<AnyWebsite>>>, Error> {
-        assert(Thread.isMainThread)
-
-        let context = self.container.viewContext
-        let request = CD_Website.request
-        request.predicate = query.cd_predicate
-        request.sortDescriptors = query.cd_sortDescriptors
-
-        let controller = NSFetchedResultsController(fetchRequest: request,
-                                                    managedObjectContext: context,
-                                                    sectionNameKeyPath: nil,
-                                                    cacheName: nil)
-        do {
-            try controller.performFetch()
-            return .success(
-                AnyListObserver(
-                    FetchedResultsControllerListObserver(controller) { [websiteCache] site in
-                        AnyElementObserver(websiteCache[site.objectID] {
-                            ManagedObjectElementObserver(site, { AnyWebsite($0) })
-                        })
-                    }
-                )
-            )
-        } catch {
-            return .failure(.read(error as NSError))
+            Website.Ident(cd_website.objectID)
         }
     }
     
-    func update(_ inputs: Set<AnyElementObserver<AnyWebsite>>, _ raw: AnyWebsite.Raw) -> Result<Void, Error> {
+    func delete(_ input: Set<Website.Ident>) -> Result<Void, Error> {
         let context = self.container.viewContext
-        let check = inputs.firstIndex { !context.validate($0.value.wrappedValue,
-                                                          as: CD_Website.classForCoder()) }
-        guard check == nil else {
-            let message = "Wrong Input Type"
-            log.emergency(message)
-            fatalError(message)
+        let coordinator = self.container.persistentStoreCoordinator
+        let cd_ids = input.compactMap {
+            coordinator.managedObjectID(forURIRepresentation: URL(string: $0.id)!)
         }
-        
-        var inputs = inputs
-        var changesMade = false
-        
-        while !inputs.isEmpty {
-            let input = inputs.popFirst()!
-            let website = input.value.wrappedValue as! CD_Website
-            
-            if let title = raw.title {
-                changesMade = true
-                website.cd_title = title
-            }
-            if let originalURL = raw.originalURL {
-                changesMade = true
-                website.cd_originalURL = originalURL
-            }
-            if let resolvedURL = raw.resolvedURL {
-                changesMade = true
-                website.cd_resolvedURL = resolvedURL
-            }
-            if let isArchived = raw.isArchived {
-                changesMade = true
-                website.cd_isArchived = isArchived
-            }
-            if let thumbnail = raw.thumbnail {
-                changesMade = true
-                website.cd_thumbnail = thumbnail
-            }
+        let cd_tags = cd_ids.compactMap {
+            context.object(with: $0) as? CD_Website
         }
-        
-        return changesMade ? context.datum_save() : .success(())
+        guard cd_tags.isEmpty == false else { return .success(()) }
+        cd_tags.forEach {
+            context.delete($0)
+        }
+        return context.datum_save()
     }
-    
-    func delete(_ inputs: Set<AnyElementObserver<AnyWebsite>>) -> Result<Void, Error> {
-        let context = self.container.viewContext
-        let check = inputs.firstIndex { !context.validate($0.value.wrappedValue,
-                                                          as: CD_Website.classForCoder()) }
-        guard check == nil else {
-            let message = "Wrong Input Type"
-            log.emergency(message)
-            fatalError(message)
-        }
-        
-        var inputs = inputs
-        var changesMade = false
-        
-        while !inputs.isEmpty {
-            let input = inputs.popFirst()!
-            let website = input.value.wrappedValue as! CD_Website
-            changesMade = true
-            context.delete(website)
-        }
 
-        return changesMade ? context.datum_save() : .success(())
-    }
-*/
     // MARK: Tag CRUD
     
     internal func createTag() -> Result<Tag.Ident, Error> {
-        return self.cd_createTag().map { Tag.Ident($0.objectID) }
-    }
-    
-    internal func cd_createTag() -> Result<CD_Tag, Error> {
         assert(Thread.isMainThread)
         let context = self.container.viewContext
         let tag = CD_Tag(context: context)
-        return context.datum_save().map { tag }
+        return context.datum_save().map { Tag.Ident(tag.objectID) }
     }
-/*
-    func readTags() -> Result<AnyListObserver<AnyRandomAccessCollection<AnyElementObserver<AnyTag>>>, Error> {
-        assert(Thread.isMainThread)
-
-        let context = self.container.viewContext
-        let request = CD_Tag.request
-        request.sortDescriptors = [.init(
-            key: #keyPath(CD_Tag.cd_name),
-            ascending: true,
-            selector: #selector(NSString.localizedCaseInsensitiveCompare(_:))
-        )]
-        let controller = NSFetchedResultsController(fetchRequest: request,
-                                                    managedObjectContext: context,
-                                                    sectionNameKeyPath: nil,
-                                                    cacheName: nil)
-        do {
-            try controller.performFetch()
-            return .success(
-                AnyListObserver(
-                    FetchedResultsControllerListObserver(controller) { [tagCache] tag in
-                        AnyElementObserver(tagCache[tag.objectID] {
-                            ManagedObjectElementObserver(tag, { AnyTag($0) })
-                        })
-                    }
-                )
-            )
-        } catch {
-            return .failure(.read(error as NSError))
-        }
-    }
-
-    func update(_ input: AnyElementObserver<AnyTag>, name: Optional<String?>) -> Result<Void, Error> {
-        let context = self.container.viewContext
-        guard context.validate(input.value.wrappedValue, as: CD_Tag.classForCoder()) else {
-            let message = "Wrong Input Type"
-            log.emergency(message)
-            fatalError(message)
-        }
-
-        let tag = input.value.wrappedValue as! CD_Tag
-        var changesMade = false
-        if let newName = name {
-            changesMade = true
-            tag.cd_name = newName
-        }
-
-        return changesMade ? context.datum_save() : .success(())
-    }
-
- */
+    
     internal func delete(_ input: Set<Tag.Ident>) -> Result<Void, Error> {
         let context = self.container.viewContext
         let coordinator = self.container.persistentStoreCoordinator
@@ -227,6 +86,7 @@ extension CD_Controller: Controller {
         let cd_tags = cd_ids.compactMap {
             context.object(with: $0) as? CD_Tag
         }
+        guard cd_tags.isEmpty == false else { return .success(()) }
         cd_tags.forEach {
             context.delete($0)
         }
